refactor: eliminate circular dependency between lib-queues and app-api
## Summary
Refactored lib-queues to be a pure utility package without app-specific dependencies, solving the circular dependency issue where lib-queues depended on apiclient types.

## Changes
- Split lib-queues into minimal core utilities (TakaroQueue, TakaroWorker, redis config)
- Moved queue data definitions to app-api/src/workers/dataDefinitions.ts
- Moved QueueService singleton to app-api/src/workers/QueueService.ts
- Created local queue configuration in app-api/src/workers/queuesConfig.ts
- Updated all imports to use local DTOs instead of apiclient
- Updated app-connector to define its own queue data types locally
- Renamed IQueuesConfig to IRedisConfig in lib-queues for clarity
- Modified getBullBoard to accept queues as parameter

## Benefits
- No more circular dependencies blocking development
- Better encapsulation - each app owns its queue definitions
- Type safety with direct DTO usage
- Cleaner architecture with lib-queues as true utility library
