+++ packages/lib-queues/src/bullboard.ts
@@ -1,17 +1,16 @@
 import { createBullBoard } from '@bull-board/api';
 import { ExpressAdapter } from '@bull-board/express';
 import { BullMQAdapter } from '@bull-board/api/bullMQAdapter.js';
-import { queueService } from './QueueService.js';
 import { TakaroQueue } from './TakaroQueue.js';
 
-export function getBullBoard() {
+export function getBullBoard(queues: TakaroQueue<any>[] = []) {
   const serverAdapter = new ExpressAdapter();
   serverAdapter.setBasePath('/queues');
 
-  const queues = Object.values(queueService.queues).map((queue) => new BullMQAdapter(queue.queue.bullQueue));
-  const reconcilerQueue = new TakaroQueue<Record<string, unknown>>('domainReconciler');
+  const adapters = queues.map((queue) => new BullMQAdapter(queue.bullQueue));
+
   createBullBoard({
-    queues: [...queues, new BullMQAdapter(reconcilerQueue.bullQueue)],
+    queues: adapters,
     serverAdapter: serverAdapter,
   });
