+++ packages/lib-queues/src/main.ts
@@ -1,8 +1,5 @@
-export { queueService } from './QueueService.js';
-export { queuesConfigSchema, IQueuesConfig } from './config.js';
-
-export * from './dataDefinitions.js';
 export { TakaroQueue } from './TakaroQueue.js';
 export { TakaroWorker } from './TakaroWorker.js';
-
 export { getBullBoard } from './bullboard.js';
+export { getRedisConnectionOptions } from './util/redisConnectionOptions.js';
+export { IRedisConfig, redisConfigSchema } from './config.js';
